+++ src/jvmgo/ch08/instructions/stores/lstore.go
@@ -3,11 +3,11 @@ package stores
 import "jvmgo/ch08/instructions/base"
 import "jvmgo/ch08/rtda"
 
-//store long into local variable
+// Store long into local variable
 type LSTORE struct{ base.Index8Instruction }
 
 func (self *LSTORE) Execute(frame *rtda.Frame) {
-	_lstore(frame, self.Index)
+	_lstore(frame, uint(self.Index))
 }
 
 type LSTORE_0 struct{ base.NoOperandsInstruction }
@@ -34,7 +34,6 @@ func (self *LSTORE_3) Execute(frame *rtda.Frame) {
 	_lstore(frame, 3)
 }
 
-//公用函数
 func _lstore(frame *rtda.Frame, index uint) {
 	val := frame.OperandStack().PopLong()
 	frame.LocalVars().SetLong(index, val)
